+++ sisl/io/siesta/siesta_nc.py
@@ -18,6 +18,7 @@ from sisl.physics import DensityMatrix, EnergyDensityMatrix
 from sisl.physics import DynamicalMatrix
 from sisl.physics import Hamiltonian
 from sisl.physics.overlap import Overlap
+from .._help import grid_reduce_indices
 from ._help import *
 try:
     from . import _siesta
@@ -330,7 +331,7 @@ class ncSileSiesta(SileCDFSiesta):
 
         return list(self.groups['GRID'].variables)
 
-    def read_grid(self, name, spin=0, **kwargs):
+    def read_grid(self, name, index=0, **kwargs):
         """ Reads a grid in the current Siesta.nc file
 
         Enables the reading and processing of the grids created by Siesta
@@ -339,13 +340,15 @@ class ncSileSiesta(SileCDFSiesta):
         ----------
         name : str
            name of the grid variable to read
-        spin : int or array_like, optional
+        index : int or array_like, optional
            the spin-index for retrieving one of the components. If a vector
            is passed it refers to the fraction per indexed component. I.e.
            ``[0.5, 0.5]`` will return sum of half the first two components.
            Default to the first component.
+        spin : optional
+           same as `index` argument. `spin` argument has precedence.
         """
-        spin = kwargs.get('index', spin)
+        index = kwargs.get("spin", index)
         geom = self.read_geometry()
 
         # Shorthand
@@ -376,15 +379,10 @@ class ncSileSiesta(SileCDFSiesta):
 
         if len(v[:].shape) == 3:
             grid.grid = v[:, :, :] * unit
-        elif isinstance(spin, Integral):
-            grid.grid = v[spin, :, :, :] * unit
+        elif isinstance(index, Integral):
+            grid.grid = v[index, :, :, :] * unit
         else:
-            if len(spin) > v.shape[0]:
-                raise SileError(f'{self}.read_grid requires spin to be an integer or '
-                                'an array of length equal to the number of spin components.')
-            grid.grid[:, :, :] = v[0, :, :, :] * (spin[0] * unit)
-            for i, scale in enumerate(spin[1:]):
-                grid.grid[:, :, :] += v[1+i, :, :, :] * (scale * unit)
+            grid_reduce_indices(v, np.array(index) * unit, axis=0, out=grid.grid)
 
         try:
             if v.unit == 'Ry':